requirements.txt
langchain
streamlit
cohere
tiktoken
python-dotenv
faiss-cpu
langchain-cohere